backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pymongo==4.7.2
motor==3.4.0
imagehash==4.3.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
numpy==1.26.4
